Rename package bhttp -> http:
HandlerMapping type in the http package
allows func main to not need the stdlib
http package and thus allowing bhttp to
be http.
